backend/requirements.txt
# Web框架
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# 数据处理（使用最新版本，有预编译包）
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0  # Excel支持

# Jupyter内核
jupyter-client>=8.6.0
ipykernel>=6.29.0
nbformat>=5.9.0

# AI模型调用
openai>=1.50.0
anthropic>=0.39.0

# 多智能体框架
crewai>=0.41.0
crewai-tools>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# 图表生成
matplotlib>=3.9.0
seaborn>=0.13.0

# 科研级可视化增强 ⭐
plotly>=5.18.0  # 交互式图表
kaleido>=0.2.1  # Plotly静态图导出
scipy>=1.11.0  # 科学计算和统计检验
scikit-learn>=1.3.0  # 机器学习（用于PCA等）
statsmodels>=0.14.0  # 统计建模和假设检验
lifelines>=0.27.0  # 生存分析（Kaplan-Meier曲线）
matplotlib-venn>=0.11.9  # 韦恩图
adjustText>=0.8  # 图表文字标注优化
pingouin>=0.5.3  # 统计分析（t-test, ANOVA等）

# 工具库
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# 文件处理
aiofiles>=24.0.0
PyPDF2>=3.0.0  # PDF文件解析
python-docx>=1.1.0  # Word文档解析

# 安全（可选）
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.0

